Add notebook for ground truth nyc res experiment
This is the initial notebook showing how to import the data, read
it from protobuffer format and then put it into a dataframe, and
show the head of the dataframe. I've already done this in a separate
script but as we're planning to show our work using Jupyter Notebook
I'm translating it over.

p.s. good idea to use Jupyter Notebook